Slicer: improve removing of unnecessary blocks
A block cannot be removed, if its removal would lead to a phi node receiving
two different values from a same basic block. Moreover, if both edges from a
basic block can be removed, we keep the one that does not lead to the exit block
(if such edge exists).