Add functional regression tests
We use py.test framework for the test generation. Each test is specified by a
folder in tests/kernel_modules. A test folder contains source files to compare
and a test specification file test.yaml that decribes the module name, the module
parameter, and the functions from the module that work with the parameter along
with expected analysis output for each function.