Slicer: improve computation of affected blocks and included successors
Affected blocks are those that are not reachable through all branches of
a terminator instruction. We compute set of reachable blocks from a single
branch by replacing the terminator by an unconditional branch and using
isPotentiallyReachable on the modified CFG.

Included successors are those, through which it is possible to reach
an included block, that cannot be reached through other successors. We use
same computation principle as for affected blocks.

These function rely on a fact that llvm::isPotentiallyReachable always
terminates, which requires a custom modification of this functions in order
to remove limit or explored blocks.