Regression tests: create separate unsliced LLVM for each parameter and link it
This is needed since the resulting LLVM is linked against other object files
that contain definitions of functions called from analysed functions. The list
of undefined functions is determined from the sliced LLVM which is different
for each parameter.
Also, unsliced file must be also linked in preparation phase, so that we can
re-run the slicer as a part of testing suite and the resulted linked and sliced
file is obtained.
Also support timeout result in spec files - such tests are not run to decrease
testing time.