Unit tests for LlvmKernelBuilder and LlvmKernelModule
Tests are written in pytest and can be found in the tests/unit_tests folder.
They test all important methods of given classes. These were not tested by
regresssion tests because they usually do not rebuild kernel modules. Also, now
the downloading and setup of the kernel is also tested.