Automatic downloading from Brew
If required kernel version contains release part (e.g. 3.10.0-655), it cannot
be found in the upstream and must be downloaded from Red Hat. Kernel sources
are distributed as RPM packages. We download the SRPM of kernel, extract files
out of it, and then extract the tar file containing kernel sources.