Builder: link dependent modules together
Link module against all other modules that it depends on. The list of modules
can be retrieved using the modinfo command.

This requires some changes to LLVM C API, so it is needed to checkout new
changes in https://github.com/viktormalik/llvm/commits/diffkemp or to download
the latest version of the development container from DockerHub
(viktormalik/diffkemp:latest).