SimpLL: provide information on which functions need to be analysed semantically.
In GlobalNumberState, assume that called functions of the same name are equal
(but still do their comparison). After all functions are compared, check if all
of them are truly equal. If not, report the set of functions that were not
equal. Then, it is sufficient that these are checked for semantical equality to
prove that the originally compared functions are semantically equivalent.