SimpLL: return not only function names but their files as well.
SimpLL now returns pairs of functions that are syntactically different. For
each file it also provides an absolute path to the file that the function is
defined in. This requires the LLVM module to be compiled with debug info.

This information is then used when printing syntactic diff of functions.