Package DiffKemp into a Docker container.
The container is built from the development environment container
diffkemp-devel.
Also provide a script to simplify running DiffKabi in this container. The
script accepts the same arguments as DiffKabi does and mounts the kernel
directories (if they exist) into the container so that the kernels do not have
to be downloaded.