kernel source: improve finding functions with definition of a symbol
Now we use cscope to find all definitions and *all* usages of the given
function and return list of found files ordered by priority (see comment in
kernel_source.py).
This is needed if a function is defined in header file only. Since we now
include all usages of the function, the given header will be included in one of
them and the function body will be built.