SimpLL: slicer: fix bug with including stores between alloca and its usage
Every time some instrution with 'alloca' as an operand is included, check all 'store'
between the alloca and the usage and include those as well.
Until now, this was only done when the alloca operand was added into the set of
included instructions for the first time.
Also add regression test that revealed this bug.