Handle changes of macro and enum values.
We use the macro debug info obtained by using the -fdebug-macro option in clang.
First, all constants having the value of some macro in the first program are
collected.
Then, we find the used macros in the other module and check if they have the
same value. If not, we create a mapping between the constant and the new value.
Finally, this information is used in the FunctionComparator.

This approach is generally unsound since we cannot be sure that some value has
truly been generated from a macro. Therefore, if the value of some macro
changed and there is some constant used (not generated from the macro) having
the same change of its value, this change will not be discovered. This is,
however, quite unlikely.