SimpLL: handle changes in basic block structure due to code moved to function.
In case functions have different numbers of BBs, they can be compared as
unequal even before comparing their instructions (e.g. because they are
assigned different serial numbers and then appear in a phi node). However, this
situation can be caused by moving part of the functionality into a function and
hence we treat the BBs in cmpValues as equal to continue the comparing and
enable trying the function inlining. This should be safe because if a
difference occurs in basic block structure, it will eventually occur in the
basic block contents.

This change requires a newer development container since serial maps had to be
made protected in the LLVM's FunctionComparator.