Removed unnecessary diffkemp tests.
Test wm97xx_ts tested GEPs, but this is already tested in
another test (b2c2-flexcop) checking a new version of cmpGEPs
(this version fixed handling structures with numbers in their
names, but obviously this also checks the cmpGEPs function as a
whole).

Test ipmi_si was for handling memory allocation outside the main
function in llreve, but the function now get compared as
syntactically equal by SimpLL, making the test useless.

Test usbcore was an old test that took a lot of time to complete.

Tests opl3, sony-laptop were checking a fix in llreve, but llreve
is not called at all in the present version, making the tests
pointless. (Similar to ipmi_si.)

Test tcp_bic was supposed to test better global value number
assignment, but this was a common error that is also checked by
tests in diffkabi.