SimpLL: improve synchronization of serial maps when comparing BBs in cmpValues
When comparing BBs in cmpValues, we treat them as equal even if they are
compared as non-equal (see commit fb7f16c).
Here, we improve synchronization of the serial maps in FunctionComaparator in
case when just of the blocks was newly inserted. Instead of adding a
placeholder to the other map to synchronize, we remove the newly added block.