SimpLL: do not inline functions already compared as non equal
If a function that was already compared as non equal is inlined, its name can
be removed by inlining, which causes a failure when trying to report
information about the function.
Also if some called function is inlined in a single module only, simplify the
function corresponding to the caller in the other module (because inlining does
the simplification).