SimpLL: report functions with missing definitions that could be inlined
If a function call that could be inlined calls a function that is not defined
in the module, report this in the SimpLL output. We create a new field in the
output YAML "missing-defs" that contains pairs of functions from the
first/second module to be linked to the module.
Also, function inlining is heavily improved. Instead of marking functions as
'alwaysinline' and running the AlwaysInliner pass, we call the InlineFunction
method that inlines a single CallInst.
Due to this, several minor changes are introduced:
 - tryInline is a pair of call instructions instead of functions
 - a function to get a non-const pointer to a CallInst from the const one