Module linking: enable restoring the unlinked version of a module
This is used in function_diff if there are some missing function definitions
that require other module to be linked.
If we didn't do this, the linked file could be reused later and hence analysing
multiple functions (e.g. KABI) in a row could give different results than
running functions individually.
Also, test this restoring in unit tests.