SimpLL: SourceCodeUtils: Retrieve and propagate macro arguments.
Adds two additional fields to MacroElement: fullName, to which the
full name fo the macro including parameters is recorded, and args,
which stores the arguments passes to the macro from its call.
Functions used for function call parsing are used to get both
parameters and arguments and findAndReplace is used to make sure
the parameters of macros are expanded according to the arguments
of the root macro.