SimpLL: Generate field access abstractions.
Add a pass that takes blocks starting with a GEP instruction and
containing GEP and cast instructions (in case they meet certain
conditions, see the comments in the code) and creates a separate
function into which the block is moved. This makes comparing the
block as a whole simple, while also retaining the possibility to
compare it instruction by instruction by inlining.

The pass also adds one of this whole-unit field access comparisons -
the comparison of such blocks by the total constant offset, which
solves a common false positive case where a field is moved into an
anonymous structure, union or both.