SimpLL: Set debug location in field accesses.
This enables the field access call to be used in the macro finding
function, fixing a bug where one of the macros is not found because
the macro scope is removed by inlining of a function call alongside
with the inlining of the field access call.

This also sets the debug location to the instructions inside the
field access abstraction body to enable it to be found in case the
macros would be only found after inlining. (The original DebugLoc
objects can't be kept, that causes an error for some reason).