SimpLL: Better function inlining.
Inlines calls in cmpOperations only in the case when the result of
that function is non equal, i.e. does not inline when the names
differ, because with the new cmpGlobalValues implementation there
may be a case when the function names are diffent, but the functions
are compared as equal, leading to a bug.

In addition inlining when a call instruction operand differs (i.e.
in cmpBasicBlocks) is added to solve the case for which the removed
inlining of calls to functions with different names was for (i.e.
when cmpOperations, which does not compare arguments, returns 0,
but the functions are different).