SimpLL: Added custom pass for function merging.
The pass merges functions that differ only in name number suffix and
have the same type together into one function without the suffix.

This fixes cases when multiple diffs between numbered versions of the
same function are passed to DiffKemp, which may show the wrong one (it
shows only one of them) and some other problems in function comparison.