Improve printing results to files
By default, diffs and callstacks are printed to files. A single file is
created for each compared symbol that is different. In case that groups
are used, there is a separate directory for each group.

All diff files (and group directories) are stored inside a directory
that can be specified from the command line.

If there are unknown or error results, these are always printed to
stdout.