SimpLL: Handle suffixes in compareFunctions.
The function compareFunctions does an additional comparison of names
when both functions are declarations and it didn't drop number suffixes,
leading to reporting a difference between a variation of a function
generated by RemoveUnusedReturnValuesPass and the original function from
the second module. This is fixed by using the names with the suffixes
dropped in the name comparisons.