SimpLL: Fix inlining causing empty diffs.
The problem was that there were some cases when the call instructions
differed in a way that didn't show up in the C source code (e.g. the
return type changed without affecting the call), leading to inlining,
but the original difference was inside the inlined function (and after
the inlining it was reported in the parent one).

The fix uses a similar technique to syntax difference handling - in case
two functions with the same name are being inlined, they are compared
using ModuleComparator and in case both the original and the inlined
functions are compared as non-equal, the difference between the original
ones is marked as covered (in case the original functions are equal after
the inlining, the difference of the inlined function is removed.