SimpLL: Fix bug in findInlineAssemblySourceArguments.
The problem was in the code that extracts inline assembly arguments
from C source code - it didn't check whether a starting bracket was
found and the value was passed directly to getSubstringToMatchingBracket.

Note: the problem didn't appear when using a standard C++ library that
does not assert on invalid string index.