SimpLL: Better call comparison.
If the call instructions have a difference that is detected before the
code get to the comparison of the functions itself, call cmpGlobalValues
to explicitly compare them and mark the parent function as covered in
case cmpGlobalValues returns zero. (This prevents false empty diffs in
cases when the difference in the parent function is actually in the
called function in the C source code.)

Also move the code doing additional comparisons on the calls to a
separate function.