SimpLL: Replace asm to value maps with metadata.
Previously each inline assembly code abstraction had a correspoding
entry in a map contaning its (function) name and its content. This
caused a bug when the content wasn't found after the function was
renamed by a pass.

To handle this, the assembly code was moved to a metadata node attached
to the abstraction function together with the constraint, which is now
also used while comparing inline assembly code abstraction calls.