SimpLL: changes in output
Results of all functions are now reported (both equal and not equal).
Also, callstacks are no longer computed in SimpLL, instead, the report
for each compared function contains a list of called functions (along
with the call location). With this information, it is possible to
reconstruct the callstack. Non-function differences (in macros, inline
assembly code, and types) are bound to the function they occur in, which
allows us to get rid of the "covered" flag in function reports.

These chages required a number of changes in the code:
- A new file Result is created that contains all classes representing
  the parts of the output.
- Code in Output.cpp is much simplified, now it only contains
  serialization of classes from Result.h into YAML.