Handle the new SimpLL output format in DiffKemp.
simplify_modules_diff now returns a graph where function differences
(with non-function differences attached to them) are vertices and
function calls are edges. The objects_to_compare list is now generated
from the graph in functions_diff. (The function was also renamed to
run_simpll since it now does much more than the simplification of the
modules.)

This enables the implementation of efficient result caching. Also
callstacks are shorter, because they are generated from the graph using
a breadth-first search algorithm.

Note: functions with unresolved missing definitions are now compared as
unknown