Implement result caching.
A new method was added to ComparisonGraph that merges a graph into
another one. This is used together with a new option of SimpLL that
enables the generation of a partial result graph (omitting some function
results) to only compare functions that haven't been compared yet and
reuse the results of the previous comparisons.