Treat function declaration comparison differently.
Instead of treating two declarations with the same name as equal and a
declaration and a definition as unknown (before linking), create a
special result type "assumed-equal" for both.

This type is treated as equal, except for the fact it is not cached, i.e.
if the functions bodies appear together in one module in the future,
they will be properly compared by SimpLL and a change in the
absorb_graph method ensures that the new vertex (with the exact result)
takes the place of the old one (containing only an assumption of
equality).

This fixes a problem that appeared after the implementation of result
caching, when results from such assumptions were cached and the function
were never compared, leading to additional false negatives in the
output (apart from reducing the number of false negatives that were
already present before the introduction of caching for a similar
reason).