Implement comparison boundaries in reachable_from.
When searching for all function differences affecting a certain KABI
symbol using depth-first search, include only those vertices whose file
is either a header file or the same file as the one of the starting
vertex (the starting vertex represents the KABI symbol).

Also a part of the function functions_diff was moved to a method of
ComparisonGraph and simplified a little bit.