Linking global constants externally defined
If there is a syntactic difference in names of constant global variables
that are declared as external, sources that define these constants are
found, linked to the current module, and the analysis rerun.

To regression tests was added test of sprintf function.

Note: Implementation of MissingDefs is changed from const *Function pair
to const *GlobalValue pair.