Caching: Implement faster call stack generation.
Do not run a breadth-first search for each function difference to get
the callstack. Instead, replace the DFS in reachable_from by a BFS
returning a map from vertices to the edges by which there were found in
the algorithm (in addition to the function list). The map can then be
used for backtracking in a linear time.