Caching: Exclude some functions from cache.
There are functions defined in header files that call a function in a
non-header source file. It is possible that the header file is included
from a different file than the one with the second function, leading to
SimpLL assuming the equality of it. When a function in the same
non-header file including the header is subsequently compared, the
cached result is used and the assumed equal function is not properly
compared.

Because of this function are detected by reversed graph search and
excluded from the cache file (re-comparing the assumed equal function
instead of excluding its parent would be a little bit faster, but it
would have the cost of having a much more complicated cache file
format).