SimpLL: optimise field access abstractions
It is useless to create field access abstractions if some GEP
instruction does not have a constant index since, in such case, we are
not able to compare the abstraction anyway.

Reducing the number of abstractions leads to a lower number of
inlinings, which is a rather important optimisation - inlining takes
time and after it is done, the current function must be re-analysed from
beginning.

Also, if both abstractions contain a single GEP instruction, it is
sufficient to use the original cmpGEPs method to compare them.