Caching: Better graph merging.
A vector with more successors now has a higher priority. This is done to
correctly include calls that appeared after linking.

Note: the case of a call that is actually important for the result
appearing after linking (i.e. such that it calls in function in the same
module) is quite rare, but it can happen when the call actually is
present before the linking and wasn't included only because there was a
difference before it (that is now solved by linking and inlining).