SimpLL: fix bug in SimplifyKernelFunctionCallsPass
Second argument of __might_sleep is not neccessarily a constant, even
though it is a line number. This is because the function can be called
from another function (e.g. __might_fault).

This commit prevents the pass to fail if the argument is not a constant.
Also, __might_fault is added to the list of simplified functions.