Fix bug in snapshot YAML parsing.
Snapshot._from_yaml checked whether the snapshot has groups or not
inside a for loop iterating over yaml_dict["list"], which contains
functions instead of groups if there is no group. Because of the use of
another for cycle for iterating over functions that is supposed to
iterate over functions inside a group, the iteration was quadratic,
doing the parsing as many times as many functions were in the snapshot.

The bug was fixed by wrapping yaml_dict["list"] in the "no group" case
into a list to ensure the first for loop only iterates over the list
once.