Cache CScope lookups.
A cache was added to the KernelSource class that remembers the results
of previous runs of CScope and uses them (avoiding calling CScope again)
when possible.

This improves the performance of the compare phase, because it allows
functions with missing definitions in multiple KABI symbols to be looked
up in the cache.