Support calling SimpLL through Python CFFI.
C interface was added to SimpLL (in files FFI.cpp and FFI.h) and support
for calling SimpLL through it using CFFI was added to DiffKemp.

The compare option --enable-simpll-ffi enables this feature. Currently it
doesn't work well with llvmcpy, so it is disabled in tests and not
enabled by default, but can be used with KABI comparisons to improve
performance.

The CFFI module was added to setup.py so it is built together with DiffKemp.
This means that SimpLL has now to be already built in the "build" directory
before installing DiffKemp since the module links the auxilliary static
library simpll-lib (libsimpll-lib.a).

In addition an option to CMake to (re)build the CFFI module while building
SimpLL (with an dependency on simpll-lib) called SIMPLL_REBUILD_BINDINGS
was added. It is not enabled by default because it needs CFFI to be
installed, which is typically done in setup.py and setup.py needs SimpLL
to be already built.