Add optional module caching.
If module caching is enabled, modules that occur frequently in the
snapshot are pre-loaded into a module cache. Later when they are
compared the cached version is used.

This feature requires the FFI bindings of SimpLL to be used for the
comparison.

Note: the cached modules still have to be cloned before the comparison,
which actually makes DiffKemp with module caching enabled slower. This
will change when smart cloning that copies only the parts that are
needed for the currently compared function is implemented.