SimpLL: fix bug in SourceCodeUtils
The function getSubstringToMatchingBracket may receive position equal to
string::npos. In such case, it must return an empty string to prevent
crash.

This situation may happen when a function is used in code without
brackets, which is, e.g., case for the function __dynamic_dev_dbg
occurring as an argument of macro _dynamic_func_call in newer kernels.