LLVM build: do not show sed warnings
There may be warnings when running sed for disabling goto features if
the altered header does not exist in the given kernel version. Therefore
we check for existence of the header before running the sed command.

Also, the sed stderr is redirected to /dev/null to prevent from other
potential warnings.