Handle comparison of inverse branch conditions
Pairs of compared cmp instructions with inverse predicates are collected
and if such instructions are used as conditions into branch
instructions, operands of one of the branch instructions are swapped.

Note: it is not sufficient to compare the branches as equal since the
order of successors is important in FunctionComparator::compare where it
is used to match corresponding basic blocks.