SimpLL: better comparison of PHI instructions
Handle comparison of PHI instructions with reordered incoming
block/value pairs. To handle this properly, PHI instructions are
compared at the end after all the other instructions and basic blocks
were compared. This is to ensure that values and blocks incoming to PHIs
are properly matched at the time of PHI comparison.

These changes require overloading of FunctionComparator::compare,
therefore it must be made virtual in FunctionComparators for all LLVM
versions.