SimpLL: unify finding of syntax differences
DifferentialFunctionComparator: Move all calls to all methods for
finding differences in macros, inline assembly code, or types into a
newly created function. The function is called from cmpBasicBlocks after
a difference is found in the compared operations (or in one of its
operands).

Moreover, introduce a function for finding type difference in chained
field access abstractions that is called from cmpBasicBlocks for each
pair of field abstractions, even if they are evaluated as equal (as was
previously done from cmpOperations).

Code for setting functions to be inlined is also moved to the same
place. Therefore, inlining is now set from cmpBasicBlocks only and hence
the CmpOperationsInlining unit test is merged into the
CmpBasicBlocksInlining test.