Handle comparison of relocated code
Support correct analysis of modules where a sequential block of code was
relocated to a different part of the compared function. This is done
using a newly introduced structure RelocationInfo in the
FunctionComparator. Currently supports relocation of a sequential block
of code (it can be relocated into a different basic block, however).

Since the relocation detection is done in cmpBasicBlocks, many unit
tests will fail since cmpBasicBlocks will return 0 when one of the
blocks has some extra code (the extra code is a possibly relocated
code). The unit tests are updated not to test results of such
cmpBasicBlocks calls.

TODO: the analysis should check that the relocated code does not depend
on any code that was skipped over by the relocation. This is partly
secured by the SSA property of LLVM IR, however, we should check that
if a pointer read occurs in the relocated code, then no writes to the
same pointer occur in the skipped code.

The change also required to make the beginCompare method virtual in all
FunctionComparators.