SimpLL: Remove code using field access abstractions.
Two new functions getFieldAccessStart and
isFollowingFieldAccessInstruction were added to SimpLL to support
iterating over instructions belonging to a field access operation
in place. Unlike FieldAccessFunctionGenerator, these functions rely on
the continuity of the operations (i.e. all instructions except for the
first one operate on the result of the previous instructions) instead
of their debug locations.

A new file FieldAccessUtils.cpp was created for the new functions,
isMemoryAccessToPtr (renamed to isConstantMemoryAccessToPtr) and
getFieldAccessSourceTypes.

Functions getFieldAccessSourceTypes and findTypeDifferences were
modified to work with the in-place field accesses using the
aforementioned functions. Also the code for type difference detection
was modified to look for the first GEP of the field access (using
getFieldAccessStart) instead of operating on field access abstractions.

In addition to these changes in DifferentialFunctionComparator, unit
tests were written for every function in FieldAccessUtils.