SimpLL: Ignore repetitive load instructions
When ignoring values, also check whether the given value is a repetitive
load variant of a previous load with no store instructions in between.
Skip these loads in favour of the previous ones to eliminate false
positives.

Change the set of ignored instructions into a map which holds
replacement values for the ignored ones.