Add build command
The build command builds a snapshot from a Makefile project using
a gcc wrapper.

The wrapper calls clang and llvm-link in addition to the original compiler
in LLVM IR emitting mode and collects the LLVM IR output, which is then used
to generate the snapshot.

The build command also supports modifying the options for calling clang,
enabling the removal of options not compatible with Clang and appending
Clang-specific ones.