Improve inverse conditions detection
Handle situations when one condition is negated via a boolean negation
instruction. Note that LLVM implements this using XOR.

Evaluates the following two change patterns as semantics-preserving:

  %2 = icmp eq %0, %1  |  %2 = icmp eq %0, %1
                       |  %3 = xor %2, true
  br %2 %T %F          |  br %3 %F %T

  %2 = icmp eq %0, %1  |  %2 = icmp ne %0, %1
                       |  %3 = xor %2, true
  br %2 %T %F          |  br %3 %T %F

The implementation uses the mechanisms for skipping instructions in
DifferentialFunctionComparator where the xor instruction is the one
skipped here (it is redirected to its operand). At the same time,
inverseConditions is updated accordingly (by adding a new pair for the
first example above or removing an existing pair for the second example
above).