Add support for LLVM 13
The PassManager(bool) constructor is no longer supported.

The sys::fs::F_None enum value was replaced with sys::fs::OF_None.

The CloneFunctionInto function now takes a finer grained CloneFunctionChangeType
instead of a boolean to specify the scope of changes.

The Dockerfile was updated to Ubuntu 20.04 (Focal), which supports LLVM 13.

Github CI was also updated to Focal.