CMake: fix removing SimpLL.cpp from simpll-lib
SimpLL.cpp must not be compiled into simpll-lib as it is used only when
using SimpLL as a standalone executable (contains main). However, the
CMake code removing it from the list of source files failed to do so as
the list contains absolute paths.

This also causes a bug which occurs when building with
BUILD_SHARED_LIBS=On as SimpLL.cpp defines LLVM CLI options and when it
was linked twice, the options got redefined.

This commit fixes the above problem by correctly removing SimpLL.cpp
from the list of source files.