Caching: avoid hitting filename length limit
Cache files are generated by joining absolute paths of the compared
modules which, in case of long paths, may hit the limit of the filename
length (255 bytes in Linux).

We resolve this by first removing the common prefix from the module
paths and only then joininig them to form the cache filename.