Eliminate false positives due to caching
In some cases, it may happen that a function is compared as non-equal
when compared inside a larger set of functions while it is compared as
equal when compared alone. This is caused by caching of results and
happens in a very specific situation when the following conditions must
occur:
- the function has multiple implementations within the project
- the implementations are either not equivalent or we are not able to
  show their equivalence (e.g. due to heavy refactoring)
- the function is compared as a callee of another function and (for some
  reason), the different implementations are used
- once the function itself is to be compared, the wrong result is taken
  from cache even if it would be compared correctly in this case

This solves the above problem by allowing to use the cached results only
if functions from the same file were compared. Otherwise, the comparison
is re-run.