Allow not saving path to source to the snapshot
Introduces a new '--no-source-dir' option.

In some cases, it makes sense to deliberately omit storing the path to
source dir in snapshot. An example is when the path would be the same
for both compared snapshots (e.g. when comparing different Git refs)
where storing the path could lead to false negative errors.