Code relocations: handle load/store dependency
If there is a load/store dependency between the relocated code and the
skipped code, the relocation must not be compared as semantically
equivalent. The load/store dependency means that there exists a store in
the relocated block writing to a pointer which is read by a load
instruction in the skipped block (or vice-versa).

For each skipped instruction (each instruction encountered while
relocation is in the "Stored" state), this checks that it is not
depending on any instruction in the relocated block.