Refactor work with source finders and source trees
Previously, instances of SourceTree and LlvmSourceFinder (or of their
appropriate sub-types) were created in the Snapshot constructor. This is
not a good approach as Snapshot constructor did a lot of stuff under the
hood (e.g. deduced type of the SourceTree based on the function type).
Also, this approach is not extensible very well for projects besides
kernel.

This commit introduces a new approach where the appropriate instances of
SourceTree and LlvmSourceFinder are created by the caller of the
Snapshot constructor (typically the "generate" function) or by the
snapshot YAML parsing function.

One important change is that from now on, the original source tree and
the snapshot tree are always represented using the same (sub)type of
SourceTree and LlvmSourceFinder. This is facilitated by the new
"clone_to_dir" method which must be implemented by each subtype of
SourceTree and of LlvmSourceFinder.

This is also a necessary step for splitting "diffkemp generate" into
multiple commands.