Get rid of `getPointerElementType` to support opaque pointers
With opaque pointers, it is not possible to retrieve the pointee type
directly from an LLVM::PointerType object. Thus, in some functions
(cmpAllocs, cmpMemset, getCSourceIdentifierType), we do not have any way
to determine the LLVM::Type of a pointee structure.

To overcome this issue, we extract the pointee type information from
debug information from LLVM 15 onward. The old functionality is preserved
for LLVM 14 and lower.

The unit tests are also adjusted to work with opaque pointers.