Specfile: handle building the CC wrapper
The "diffkemp build" command uses a custom CC wrapper to build the
analysed project. The wrapper can be used directly as a Python script or
built into a native binary using RPython. Since the latter is more
efficient, we use it in the RPM build.

The problem is that RPython supports Python2 only, which is not
supported on Fedora anymore, so RPython and some required packages (Py)
must be downloaded manually.

Also, since the wrapper script supports both Python2 and Python3, its
shebang must be fixed after it is installed onto the system.