Dockerfile for RPM testing
Allows to create a container image with a specific version of Fedora
containing all DiffKemp dependencies and two kernels prepared for
comparison.

When the container is run, it installs DiffKemp from RPM and performs
build-kernel and compare commands.

Supports installation from Copr (default) and from a local RPM file (by
mounting the RPM into the container and setting the DIFFKEMP_RPM env
variable). See comment in the Dockerfile for usage.