Viewer: do not build in CMake by default
Doing the production build every time when the user runs CMake is quite
impractical. In addition, production build is not necessary since it is
possible to run a development server with `diffkemp view --devel`.

So, update CMake not to create the production build by default. Also,
run `npm install` when running the development server to make sure that
all the necessary packages are installed.