Add support for building with Nix
This allows to run `nix build` and obtain DiffKemp as a Nix package
without any necessary dependencies installed on the local system. It is
also the first step towards Nix-based development environment which
should eventually replace the `diffkemp-devel` container image.

So far, the build is against LLVM 16, support for multiple LLVM versions
will follow.

The build itself is pretty straightforward as Nix has a lot of automatic
steps in place. The only thing is that we're mixing C++ and Python
sources so we need to manually define buildPhase and installPhase to
include phases for both languages.

Note: The build is based on Nix flakes which are marked as experimental,
although they are now the recommended way of using Nix and it's widely
expected that the interface won't change. It's still necessary to enable
them explicitly by running:

    $ mkdir -p ~/.config/nix
    $ echo 'experimental-features = nix-command flakes' >> ~/.config/nix/nix.conf