Nix development env for building test kernels
Add new dev shell containing everything necessary to download and
prepare kernels used in tests. Inherits most of the packages from the
default dev shell but uses specific versions of some packages necessary
for building RHEL7 kernels:
- gcc 7
- make 3.81

The environment can be run via `nix develop .#test-kernel-buildenv`. It
should be sufficient to use this env to download the kernels (with
rhel-kernel-get) and the tests can be run in the default dev shell.