Add a pattern which matches reordered binary operations
During compilation to LLVM IR, binary operations which are both
commutative and associative may sometimes get reordered. This commit
enables effective matching of such reordered instructions.

Simply explained, when the operands differ in a compared reorderable
binary operation and all its users are operations of the same type,
the binary operation is skipped. The complete set of operands is then
matched during the comparison of the user.